move TableMigrator to the SPI package
I must have just made a mistake when I created this interface
