+++ hibernate-core/src/main/java/org/hibernate/tool/schema/internal/StandardTableMigrator.java
@@ -16,6 +16,7 @@ import org.hibernate.mapping.Column;
 import org.hibernate.mapping.Table;
 import org.hibernate.tool.schema.extract.spi.ColumnInformation;
 import org.hibernate.tool.schema.extract.spi.TableInformation;
+import org.hibernate.tool.schema.spi.TableMigrator;
 import org.jboss.logging.Logger;
 
 import java.util.ArrayList;
