+++ hibernate-core/src/main/java/org/hibernate/tool/schema/spi/TableMigrator.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.tool.schema.internal;
+package org.hibernate.tool.schema.spi;
 
 import org.hibernate.boot.Metadata;
 import org.hibernate.boot.model.relational.SqlStringGenerationContext;
